Simplify construction of native classes rust-side without [[Constructor]] args
It's likely that a lot of the time there'll be no [[Constructor]] behaviour
taking args when constructing a native class rust-side from an existing struct
instance.

To assist with this, provide an additional method that needs no args provided.
And rename such that the new method has the shorter name of the two.